v2.0 增加登录，商品查询等功能

+++ src/main/java/miaosha/dao/UserDao.java
@@ -16,7 +16,7 @@ import miaosha.dao.domain.User;
 public interface UserDao {
 
 	@Select("select * from user where id=#{id}")
-	public User findUserById(@Param("id") int id);
+	public User findUserById(@Param("id") Long id);
 	
 	/**
 	 * 使用输入bean的属性作为输入参数
